requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
anthropic==0.25.0
python-docx==1.1.2
openpyxl==3.1.2
pandas==2.2.2
APScheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
flask==3.0.3
